Update docker.yaml workflow
This commit updates the docker.yaml workflow file. It modifies the 'command' and 'image' properties in the 'docker/scout-action' step. The 'command' property now includes 'cves' and 'recommendations' if the event is not a tag. The 'image' property is updated to use 'jkreileder/cf-ips-to-hcloud-fw' instead of 'image://jkreileder/cf-ips-to-hcloud-fw'. The 'sarif-file', 'organization', and 'to' properties remain unchanged.

Signed-off-by: Jürgen Kreileder <jk@blackdown.de>