Added OpenSSF Scorecard badge to README
Introduced the OpenSSF Scorecard badge to the project's README for
enhanced visibility into the security posture and quality of the
project's codebase. This addition aligns with ongoing efforts to
emphasize security and code health, offering users and potential
contributors immediate insight into the project's adherence to best
practices and standards in open-source security.

Signed-off-by: Jürgen Kreileder <jk@blackdown.de>